Refactor OperatorContexts
