+++ src/dataflow/operator.rs
@@ -1,11 +1,11 @@
-use std::{marker::PhantomData, slice::Iter};
+use std::slice::Iter;
 
 use serde::Deserialize;
 
 use crate::{
     dataflow::{
-        deadlines::Deadline, stream::StreamId, AppendableStateT, Data, ReadStream, State, StateT,
-        Timestamp, WriteStream,
+        context::*, deadlines::Deadline, stream::StreamId, AppendableStateT, Data, ReadStream,
+        State, StateT, WriteStream,
     },
     node::NodeId,
     OperatorId,
@@ -63,45 +63,6 @@ pub trait ParallelSink<S: AppendableStateT<U>, T: Data, U>: Send + Sync {
     fn on_watermark(&self, ctx: &mut ParallelSinkContext<S, U>);
 }
 
-/// A context structure made available to the callbacks of a `ParallelSink` operator. The context
-/// provides access to the current timestamp for which the callback is invoked along with the state
-/// of the operator.
-pub struct ParallelSinkContext<'a, S: AppendableStateT<T>, T> {
-    timestamp: Timestamp,
-    config: OperatorConfig,
-    state: &'a S,
-    phantomdata_t: PhantomData<T>,
-}
-
-impl<'a, S, T> ParallelSinkContext<'a, S, T>
-where
-    S: 'static + AppendableStateT<T>,
-{
-    pub fn new(timestamp: Timestamp, config: OperatorConfig, state: &'a S) -> Self {
-        Self {
-            timestamp,
-            config,
-            state,
-            phantomdata_t: PhantomData,
-        }
-    }
-
-    /// Get the timestamp for which the callback was invoked.
-    pub fn get_timestamp(&self) -> &Timestamp {
-        &self.timestamp
-    }
-
-    /// Get the configuration of the operator.
-    pub fn get_operator_config(&self) -> &OperatorConfig {
-        &self.config
-    }
-
-    /// Get the state attached to the operator.
-    pub fn get_state(&self) -> &S {
-        &self.state
-    }
-}
-
 /*************************************************************************************************
  * Sink: receives data with type T, and enables message callbacks to have mutable access to the  *
  * operator state, but all callbacks are sequentialized.                                         *
@@ -123,43 +84,6 @@ pub trait Sink<S: StateT, T: Data>: Send + Sync {
     fn on_data(&mut self, ctx: &mut SinkContext<S>, data: &T);
 
     fn on_watermark(&mut self, ctx: &mut SinkContext<S>);
-}
-
-/// A context structure made available to the callbacks of a `Sink` operator. The context provides
-/// access to the current timestamp for which the callback is invoked along with the state
-/// of the operator.
-pub struct SinkContext<'a, S: StateT> {
-    timestamp: Timestamp,
-    config: OperatorConfig,
-    state: &'a mut S,
-}
-
-impl<'a, S> SinkContext<'a, S>
-where
-    S: StateT,
-{
-    pub fn new(timestamp: Timestamp, config: OperatorConfig, state: &'a mut S) -> Self {
-        Self {
-            timestamp,
-            config,
-            state,
-        }
-    }
-
-    /// Get the timestamp for which the callback was invoked.
-    pub fn get_timestamp(&self) -> &Timestamp {
-        &self.timestamp
-    }
-
-    /// Get the configuration of the operator.
-    pub fn get_operator_config(&self) -> &OperatorConfig {
-        &self.config
-    }
-
-    /// Get the state attached to the operator.
-    pub fn get_state(&mut self) -> &mut S {
-        &mut self.state
-    }
 }
 
 /*************************************************************************************************
@@ -194,62 +118,6 @@ where
     fn on_watermark(&self, ctx: &mut ParallelOneInOneOutContext<S, U, V>);
 }
 
-/// A context structure made available to the callbacks of a `ParallelOneInOneOut` operator. The
-/// context provides access to the current timestamp for which the callback is invoked along with
-/// the state of the operator and the write stream to send the outputs on.
-pub struct ParallelOneInOneOutContext<'a, S, T, U>
-where
-    S: AppendableStateT<U>,
-    T: Data + for<'b> Deserialize<'b>,
-{
-    timestamp: Timestamp,
-    config: OperatorConfig,
-    state: &'a S,
-    write_stream: WriteStream<T>,
-    phantom_u: PhantomData<U>,
-}
-
-impl<'a, S, T, U> ParallelOneInOneOutContext<'a, S, T, U>
-where
-    S: AppendableStateT<U>,
-    T: Data + for<'b> Deserialize<'b>,
-{
-    pub fn new(
-        timestamp: Timestamp,
-        config: OperatorConfig,
-        state: &'a S,
-        write_stream: WriteStream<T>,
-    ) -> Self {
-        Self {
-            timestamp,
-            config,
-            state,
-            write_stream,
-            phantom_u: PhantomData,
-        }
-    }
-
-    /// Get the timestamp for which the callback was invoked.
-    pub fn get_timestamp(&self) -> &Timestamp {
-        &self.timestamp
-    }
-
-    /// Get the configuration of the operator.
-    pub fn get_operator_config(&self) -> &OperatorConfig {
-        &self.config
-    }
-
-    /// Get the state attached to the operator.
-    pub fn get_state(&self) -> &S {
-        &self.state
-    }
-
-    /// Get the write stream to send the output on.
-    pub fn get_write_stream(&mut self) -> &mut WriteStream<T> {
-        &mut self.write_stream
-    }
-}
-
 /**************************************************************************************************
  * OneInOneOut: Receives data with type T, and enables message callbacks to have mutable access   *
  * to the operator state, but all callbacks are sequentialized.                                   *
@@ -276,60 +144,6 @@ where
     fn on_data(&mut self, ctx: &mut OneInOneOutContext<S, U>, data: &T);
 
     fn on_watermark(&mut self, ctx: &mut OneInOneOutContext<S, U>);
-}
-
-/// A context structure made available to the callbacks of a `OneInOneOut` operator. The context
-/// provides access to the current timestamp for which the callback is invoked along with the
-/// state of the operator and the write stream to send the outputs on.
-pub struct OneInOneOutContext<'a, S, T>
-where
-    S: StateT,
-    T: Data + for<'b> Deserialize<'b>,
-{
-    timestamp: Timestamp,
-    config: OperatorConfig,
-    state: &'a mut S,
-    write_stream: WriteStream<T>,
-}
-
-impl<'a, S, T> OneInOneOutContext<'a, S, T>
-where
-    S: StateT,
-    T: Data + for<'b> Deserialize<'b>,
-{
-    pub fn new(
-        timestamp: Timestamp,
-        config: OperatorConfig,
-        state: &'a mut S,
-        write_stream: WriteStream<T>,
-    ) -> Self {
-        Self {
-            timestamp,
-            config,
-            state,
-            write_stream,
-        }
-    }
-
-    /// Get the timestamp for which the callback was invoked.
-    pub fn get_timestamp(&self) -> &Timestamp {
-        &self.timestamp
-    }
-
-    /// Get the configuration of the operator.
-    pub fn get_operator_config(&self) -> &OperatorConfig {
-        &self.config
-    }
-
-    /// Get the state attached to the operator.
-    pub fn get_state(&mut self) -> &mut S {
-        &mut self.state
-    }
-
-    /// Get the write stream to send the output on.
-    pub fn get_write_stream(&mut self) -> &mut WriteStream<T> {
-        &mut self.write_stream
-    }
 }
 
 pub struct OneInOneOutSetupContext {
@@ -401,62 +215,6 @@ where
     fn on_watermark(&self, ctx: &mut ParallelTwoInOneOutContext<S, V, W>);
 }
 
-/// A context structure made available to the callbacks of a `ParallelTwoInOneOut` operator. The
-/// context provides access to the current timestamp for which the callback is invoked along with
-/// the state of the operator and the write stream to send the outputs on.
-pub struct ParallelTwoInOneOutContext<'a, S, T, U>
-where
-    S: AppendableStateT<U>,
-    T: Data + for<'b> Deserialize<'b>,
-{
-    timestamp: Timestamp,
-    config: OperatorConfig,
-    state: &'a S,
-    write_stream: WriteStream<T>,
-    phantom_u: PhantomData<U>,
-}
-
-impl<'a, S, T, U> ParallelTwoInOneOutContext<'a, S, T, U>
-where
-    S: AppendableStateT<U>,
-    T: Data + for<'b> Deserialize<'b>,
-{
-    pub fn new(
-        timestamp: Timestamp,
-        config: OperatorConfig,
-        state: &'a S,
-        write_stream: WriteStream<T>,
-    ) -> Self {
-        Self {
-            timestamp,
-            config,
-            state,
-            write_stream,
-            phantom_u: PhantomData,
-        }
-    }
-
-    /// Get the timestamp for which the callback was invoked.
-    pub fn get_timestamp(&self) -> &Timestamp {
-        &self.timestamp
-    }
-
-    /// Get the configuration of the operator.
-    pub fn get_operator_config(&self) -> &OperatorConfig {
-        &self.config
-    }
-
-    /// Get the state attached to the operator.
-    pub fn get_state(&self) -> &S {
-        &self.state
-    }
-
-    /// Get the write stream to send the output on.
-    pub fn get_write_stream(&mut self) -> &mut WriteStream<T> {
-        &mut self.write_stream
-    }
-}
-
 /**************************************************************************************************
  * TwoInOneOut: Receives data with type T and U, and enables message callbacks to have mutable    *
  * access to the operator state, but all callbacks are sequentialized.                            *
@@ -492,60 +250,6 @@ where
     fn on_right_data(&mut self, ctx: &mut TwoInOneOutContext<S, V>, data: &U);
 
     fn on_watermark(&mut self, ctx: &mut TwoInOneOutContext<S, V>);
-}
-
-/// A context structure made available to the callbacks of a `TwoInOneOut` operator. The context
-/// provides access to the current timestamp for which the callback is invoked along with the
-/// state of the operator and the write stream to send the outputs on.
-pub struct TwoInOneOutContext<'a, S, T>
-where
-    S: StateT,
-    T: Data + for<'b> Deserialize<'b>,
-{
-    timestamp: Timestamp,
-    config: OperatorConfig,
-    state: &'a mut S,
-    write_stream: WriteStream<T>,
-}
-
-impl<'a, S, T> TwoInOneOutContext<'a, S, T>
-where
-    S: StateT,
-    T: Data + for<'b> Deserialize<'b>,
-{
-    pub fn new(
-        timestamp: Timestamp,
-        config: OperatorConfig,
-        state: &'a mut S,
-        write_stream: WriteStream<T>,
-    ) -> Self {
-        Self {
-            timestamp,
-            config,
-            state,
-            write_stream,
-        }
-    }
-
-    /// Get the timestamp for which the callback was invoked.
-    pub fn get_timestamp(&self) -> &Timestamp {
-        &self.timestamp
-    }
-
-    /// Get the configuration of the operator.
-    pub fn get_operator_config(&self) -> &OperatorConfig {
-        &self.config
-    }
-
-    /// Get the state attached to the operator.
-    pub fn get_state(&mut self) -> &mut S {
-        &mut self.state
-    }
-
-    /// Get the write stream to send the output on.
-    pub fn get_write_stream(&mut self) -> &mut WriteStream<T> {
-        &mut self.write_stream
-    }
 }
 
 /*************************************************************************************************
@@ -587,72 +291,6 @@ where
     fn on_watermark(&self, ctx: &mut ParallelOneInTwoOutContext<S, U, V, W>);
 }
 
-/// A context structure made available to the callbacks of a `ParallelOneInTwoOut` operator. The
-/// context provides access to the current timestamp for which the callback is invoked along with
-/// the state of the operator and the write streams to send the outputs on.
-pub struct ParallelOneInTwoOutContext<'a, S, T, U, V>
-where
-    S: AppendableStateT<V>,
-    T: Data + for<'b> Deserialize<'b>,
-    U: Data + for<'b> Deserialize<'b>,
-{
-    timestamp: Timestamp,
-    config: OperatorConfig,
-    state: &'a S,
-    left_write_stream: WriteStream<T>,
-    right_write_stream: WriteStream<U>,
-    phantom_v: PhantomData<V>,
-}
-
-impl<'a, S, T, U, V> ParallelOneInTwoOutContext<'a, S, T, U, V>
-where
-    S: AppendableStateT<V>,
-    T: Data + for<'b> Deserialize<'b>,
-    U: Data + for<'b> Deserialize<'b>,
-{
-    pub fn new(
-        timestamp: Timestamp,
-        config: OperatorConfig,
-        state: &'a S,
-        left_write_stream: WriteStream<T>,
-        right_write_stream: WriteStream<U>,
-    ) -> Self {
-        Self {
-            timestamp,
-            config,
-            state,
-            left_write_stream,
-            right_write_stream,
-            phantom_v: PhantomData,
-        }
-    }
-
-    /// Get the timestamp for which the callback was invoked.
-    pub fn get_timestamp(&self) -> &Timestamp {
-        &self.timestamp
-    }
-
-    /// Get the configuration of the operator.
-    pub fn get_operator_config(&self) -> &OperatorConfig {
-        &self.config
-    }
-
-    /// Get the state attached to the operator.
-    pub fn get_state(&self) -> &S {
-        &self.state
-    }
-
-    /// Get the left write stream to send the output on.
-    pub fn get_left_write_stream(&mut self) -> &mut WriteStream<T> {
-        &mut self.left_write_stream
-    }
-
-    /// Get the right write stream to send the output on.
-    pub fn get_right_write_stream(&mut self) -> &mut WriteStream<U> {
-        &mut self.right_write_stream
-    }
-}
-
 /**************************************************************************************************
  * OneInTwoOut: Receives data with type T, and enables message callbacks to have mutable          *
  * access to the operator state, but all callbacks are sequentialized.                            *
@@ -686,70 +324,6 @@ where
     fn on_data(&mut self, ctx: &mut OneInTwoOutContext<S, U, V>, data: &T);
 
     fn on_watermark(&mut self, ctx: &mut OneInTwoOutContext<S, U, V>);
-}
-
-/// A context structure made available to the callbacks of a `OneInTwoOut` operator. The context
-/// provides access to the current timestamp for which the callback is invoked along with the
-/// state of the operator and the write streams to send the outputs on.
-pub struct OneInTwoOutContext<'a, S, T, U>
-where
-    S: StateT,
-    T: Data + for<'b> Deserialize<'b>,
-    U: Data + for<'b> Deserialize<'b>,
-{
-    timestamp: Timestamp,
-    config: OperatorConfig,
-    state: &'a mut S,
-    left_write_stream: WriteStream<T>,
-    right_write_stream: WriteStream<U>,
-}
-
-impl<'a, S, T, U> OneInTwoOutContext<'a, S, T, U>
-where
-    S: StateT,
-    T: Data + for<'b> Deserialize<'b>,
-    U: Data + for<'b> Deserialize<'b>,
-{
-    pub fn new(
-        timestamp: Timestamp,
-        config: OperatorConfig,
-        state: &'a mut S,
-        left_write_stream: WriteStream<T>,
-        right_write_stream: WriteStream<U>,
-    ) -> Self {
-        Self {
-            timestamp,
-            config,
-            state,
-            left_write_stream,
-            right_write_stream,
-        }
-    }
-
-    /// Get the timestamp for which the callback was invoked.
-    pub fn get_timestamp(&self) -> &Timestamp {
-        &self.timestamp
-    }
-
-    /// Get the configuration of the operator.
-    pub fn get_operator_config(&self) -> &OperatorConfig {
-        &self.config
-    }
-
-    /// Get the state attached to the operator.
-    pub fn get_state(&mut self) -> &mut S {
-        &mut self.state
-    }
-
-    /// Get the left write stream to send the output on.
-    pub fn get_left_write_stream(&mut self) -> &mut WriteStream<T> {
-        &mut self.left_write_stream
-    }
-
-    /// Get the right write stream to send the output on.
-    pub fn get_right_write_stream(&mut self) -> &mut WriteStream<U> {
-        &mut self.right_write_stream
-    }
 }
 
 #[derive(Clone)]
